backend/requirements/common.txt
flask==1.0.2
flask-cors==3.0.7
py-postgresql==1.2.1
python-dotenv>=0.10
